Created General Patient Care - Adult Protocol
Not finished, but made an avpu_checker to assess for mental status.
Added a cardiac_presence check as per early steps of the respective protocol.

The rest of the protocol isn't near finished. Need to eventually embed protocol modules into this one.
I.e. the cardiac arrest one, right now it just says do CPR.

Figure out a way to reference that (eventually) made module, or how to organize and go about jumping there

Need to figure out an organized logic to reference all the different protocols
Otherwise main.py is going to be very confusing.